Continued merging code + added dims
Continued to merge our code to work together. Currently it seems that the code fails for 2 reasons:
1 - it can't handle batches.
2. (when using batch=1) the kernel size is bigger than the input size

Also, added the option to choose the number of dimensions of our data. Our data is 3D, so in order to take 2D data we still need to implement the 2D data sampling.